AP-183: Improve scripts for non-TLS Docker deployments
Add a flag, disable_swarm_tls, to the clusters API. Use this in the
deploy-local.sh script.

Use existing REGISTRY_USERNAME, REGISTRY_PASSWORD variables if
applicable in sensor-deploy.sh.

Merge AP-183-non-tls-swarm-deploy-cg (pull request #138)
Approved-by: Connor Gorman <cgorman@stackrox.com>

+++ central/clusters/swarm.go
@@ -1,6 +1,7 @@
 package clusters
 
 import (
+	"strconv"
 	"text/template"
 
 	"bitbucket.org/stack-rox/apollo/generated/api/v1"
@@ -13,9 +14,14 @@ func init() {
 
 func newSwarm() deployer {
 	return &basicDeployer{
-		deploy: template.Must(template.New("swarm").Parse(swarmDeploy)),
-		cmd:    template.Must(template.New("swarm").Parse(swarmCmd)),
+		deploy:    template.Must(template.New("swarm").Parse(swarmDeploy)),
+		cmd:       template.Must(template.New("swarm").Parse(swarmCmd)),
+		addFields: addSwarmFields,
 	}
+}
+
+func addSwarmFields(c Wrap, fields map[string]string) {
+	fields["DisableSwarmTLS"] = strconv.FormatBool(c.DisableSwarmTls)
 }
 
 var (
@@ -40,9 +46,11 @@ services:
       - "{{.ClusterNameEnv}}={{.ClusterName}}"
       - "{{.AdvertisedEndpointEnv}}={{.AdvertisedEndpoint}}"
       - "{{.ImageEnv}}={{.Image}}"
+{{ if ne .DisableSwarmTLS "true" }}
       - "DOCKER_CERT_PATH=/run/secrets/stackrox.io/docker/"
       - "DOCKER_HOST=$DOCKER_HOST"
       - "DOCKER_TLS_VERIFY=$DOCKER_TLS_VERIFY"
+{{ end }}
     secrets:
       - source: sensor_certificate
         target: stackrox.io/cert.pem
@@ -53,6 +61,7 @@ services:
       - source: central_certificate
         target: stackrox.io/ca.pem
         mode: 400
+{{ if ne .DisableSwarmTLS "true" }}
       - source: docker_client_ca_pem
         target: stackrox.io/docker/ca.pem
         mode: 400
@@ -62,6 +71,7 @@ services:
       - source: docker_client_key_pem
         target: stackrox.io/docker/key.pem
         mode: 400
+{{ end }}
       - source: registry_auth
         target: stackrox.io/registry_auth
         mode: 400
@@ -76,38 +86,50 @@ secrets:
     file: sensor-cert.pem
   central_certificate:
     file: central-ca.pem
+{{ if ne .DisableSwarmTLS "true" }}
   docker_client_ca_pem:
     file: docker-ca.pem
   docker_client_cert_pem:
     file: docker-cert.pem
   docker_client_key_pem:
     file: docker-key.pem
+{{ end }}
   registry_auth:
     file: registry-auth
 `
 
-	swarmCmd = commandPrefix + `set -u
-WD=$(pwd)
+	swarmCmd = commandPrefix + `WD=$(pwd)
 cd $DIR
 
-echo -n "Registry username for StackRox Mitigate image: "
-read -s REGISTRY_USERNAME
-echo
-echo -n "Registry password for StackRox Mitigate image: "
-read -s REGISTRY_PASSWORD
-echo
-
+# Create registry-auth secret, used to pull the benchmark image.
+if [ -z "$REGISTRY_USERNAME" ]; then
+  echo -n "Registry username for StackRox Mitigate image: "
+  read -s REGISTRY_USERNAME
+  echo
+fi
+if [ -z "$REGISTRY_PASSWORD" ]; then
+  echo -n "Registry password for StackRox Mitigate image: "
+  read -s REGISTRY_PASSWORD
+  echo
+fi
 REGISTRY_AUTH="{\"username\": \"$REGISTRY_USERNAME\", \"password\": \"$REGISTRY_PASSWORD\"}"
 echo -n "$REGISTRY_AUTH" | base64 | tr -- '+=/' '-_~' > registry-auth
 
-cp $DOCKER_CERT_PATH/ca.pem ./docker-ca.pem
-cp $DOCKER_CERT_PATH/key.pem ./docker-key.pem
-cp $DOCKER_CERT_PATH/cert.pem ./docker-cert.pem
+# Gather client cert bundle if it is present.
+if [ -n "$DOCKER_CERT_PATH" ]; then
+  cp $DOCKER_CERT_PATH/ca.pem ./docker-ca.pem
+  cp $DOCKER_CERT_PATH/key.pem ./docker-key.pem
+  cp $DOCKER_CERT_PATH/cert.pem ./docker-cert.pem
+fi
 
+# Deploy.
 docker stack deploy -c ./sensor-deploy.yaml mitigate --with-registry-auth
 
-rm ./docker-*
+# Clean up temporary files.
 rm registry-auth
+if [ -n "$DOCKER_CERT_PATH" ]; then
+  rm ./docker-*
+fi
 
 cd $WD
 `
